refactor and work on saving timezones
